style 폴더 변경, font 추가
