+++ web/next.config.js
@@ -5,7 +5,7 @@ const nextConfig = {
   reactStrictMode: true,
   sassOptions: {
     includePaths: [path.join(__dirname, 'styles')],
-    prependData: `@import "styles/_variables.scss"; @import "styles/_mixins.scss";`,
+    prependData: `@import "styles/abstracts/_variables.scss"; @import "styles/abstracts/_mixins.scss"; @import "styles/base/_reset.scss";`,
   },
 };
