+++ web/src/components/calendar/BaroMeterDate.tsx
@@ -1,6 +1,6 @@
 import React, {useMemo} from 'react';
 import classNames from 'classnames/bind';
-import scss from 'styles/barometerDate.module.scss';
+import scss from 'styles/components/barometerDate.module.scss';
 import Image from 'next/image';
 
 const cn = classNames.bind(scss);
