+++ web/src/components/calendar/MonthlyCalendar.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useMemo, useRef, useState} from 'react';
 import classNames from 'classnames/bind';
-import scss from 'styles/calendar.module.scss';
+import scss from 'styles/components/calendar.module.scss';
 import dayjs from 'dayjs';
 import DayHeader from './DayHeader';
 import Weekly from './Weekly';
