+++ web/src/markup/components/Calendar.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+interface CalendarProps {}
+
+const Calendar = ({}: CalendarProps) => {
+  return <>
+  <div></div>
+  </>;
+};
+
+export default Calendar;
