--- web/src/markup/components/Test.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-interface TestViewProps {}
-
-const TestView = ({}: TestViewProps) => {
-  return <></>;
-};
-
-export default TestView;
