+++ web/src/stories/Markup/pages/Test.stories.ts
@@ -1,0 +1,19 @@
+import type {Meta, StoryObj} from '@storybook/react';
+
+import Test from 'markup/pages/Test';
+
+const meta = {
+  title: 'Markup/Pages/Test',
+  component: Test,
+  parameters: {
+    layout: 'centered',
+  },
+  argTypes: {},
+} satisfies Meta<typeof Test>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Primary: Story = {
+  args: {},
+};
