+++ web/src/styles/abstracts/_font.scss
@@ -1,0 +1,20 @@
+@font-face {
+  font-family: 'Montserrat';
+  font-weight: 300;
+  src: local("Montserrat"), url(../../../public/fonts/Montserrat-Light.woff2) format('woff2');
+}
+@font-face {
+  font-family: 'Montserrat';
+  font-weight: 400;
+  src: local("Montserrat"), url(../../../public/fonts/Montserrat-Regular.woff2) format('woff2');
+}
+@font-face {
+  font-family: 'Montserrat';
+  font-weight: 700;
+  src: local("Montserrat"), url(../../../public/fonts/Montserrat-Bold.woff2) format('woff2');
+}
+@font-face {
+  font-family: 'Montserrat';
+  font-weight: 900;
+  src: local("Montserrat"), url(../../../public/fonts/Montserrat-Black.woff2) format('woff2');
+}
