+++ web/src/styles/abstracts/_variables.scss
@@ -1,0 +1,47 @@
+:root {
+  --color-system-blue-blue0: rgba(37, 56, 82, 1);
+  --color-system-blue-blue10: rgba(41, 84, 140, 1);
+  --color-system-blue-blue20: rgba(54, 111, 188, 1);
+  --color-system-blue-blue60: rgba(102, 169, 241, 1);
+  --color-system-gray-gray0: rgba(15, 15, 15, 1);
+  --color-system-gray-gray10: rgba(41, 43, 46, 1);
+  --color-system-gray-gray20: rgba(52, 54, 56, 1);
+  --color-system-gray-gray30: rgba(63, 67, 71, 1);
+  --color-system-gray-gray40: rgba(92, 97, 102, 1);
+  --color-system-gray-gray5: rgba(29, 30, 31, 1);
+  --color-system-gray-gray50: rgba(120, 124, 128, 1);
+  --color-system-gray-gray60: rgba(151, 154, 158, 1);
+  --color-system-gray-gray70: rgba(181, 184, 189, 1);
+  --color-system-gray-gray80: rgba(195, 196, 199, 1);
+  --color-system-gray-gray90: rgba(206, 207, 209, 1);
+  --color-system-gray-gray95: rgba(228, 229, 229, 1);
+  --color-system-gray-gray99: rgba(250, 250, 250, 1);
+  --color-system-indigo-indigo20: rgba(66, 90, 211, 1);
+  --color-system-indigo-indigo50: rgba(99, 123, 242, 1);
+  --color-system-indigo-indigo60: rgba(121, 142, 245, 1);
+  --color-system-lime-lime40: rgba(146, 199, 68, 1);
+  --color-system-lime-lime60: rgba(181, 225, 100, 1);
+  --color-system-orange-orange40: rgba(237, 133, 55, 1);
+  --color-system-orange-orange60: rgba(243, 173, 95, 1);
+  --color-system-pink-pink60: rgba(247, 131, 172, 1);
+  --color-system-purple-purple0: rgba(66, 57, 94, 1);
+  --color-system-purple-purple20: rgba(103, 65, 217, 1);
+  --color-system-purple-purple40: rgba(121, 80, 242, 1);
+  --color-system-purple-purple60: rgba(151, 117, 250, 1);
+  --color-system-purple-purple70: rgba(177, 151, 252, 1);
+  --color-system-red-red40: rgba(231, 95, 89, 1);
+  --color-system-red-red50: rgba(237, 116, 112, 1);
+  --color-system-red-red60: rgba(239, 141, 138, 1);
+  --color-system-teal-teal20: rgba(65, 144, 107, 1);
+  --color-system-teal-teal60: rgba(110, 214, 172, 1);
+  --color-system-teal-teal70: rgba(135, 227, 192, 1);
+  --color-system-yellow-yellow0: rgba(102, 93, 66, 1);
+  --color-system-yellow-yellow20: rgba(191, 162, 80, 1);
+  --color-system-yellow-yellow60: rgba(248, 214, 91, 1);
+  --color-system-yellow-yellow90: rgba(253, 243, 197, 1);
+  --surface-contents-bg: rgba(26, 26, 26, 1);
+  --surface-surface: rgba(20, 20, 20, 1);
+  --surface-text-field: rgba(23, 23, 23, 1);
+}
+
+$font-family: "Montserrat", system-ui, -apple-system, roboto, "Apple SD Gothic Neo", sans-serif;
