+++ web/src/styles/components/barometerDate.module.scss
@@ -24,8 +24,7 @@
   }
 
   .text-wrapper {
-    color: $color-system-gray-gray99;
-    font-family: 'Montserrat Alternates-SemiBold', Helvetica;
+    color: var(--color-system-gray-gray99);
     font-size: 12px;
     font-weight: 600;
     width: calc(100% - 2px);
@@ -47,7 +46,7 @@
   }
 
   .ellipse {
-    background-color: $color-system-gray-gray5;
+    background-color: var(--color-system-gray-gray5);
     border-radius: 1.5px;
     height: 3px;
     position: relative;
