+++ web/src/styles/components/cheerup.module.scss
@@ -1,6 +1,6 @@
 .frame {
   align-items: center;
-  background-color: $color-system-gray-gray0;
+  background-color: var(--color-system-gray-gray0);
   border-radius: 12px;
   display: flex;
   gap: 8px;
@@ -10,8 +10,7 @@
 }
 
 .frame .text-wrapper {
-  color: $color-system-gray-gray60;
-  font-family: 'Montserrat Alternates-Bold', Helvetica;
+  color: var(--color-system-gray-gray60);
   font-size: 12px;
   font-weight: 700;
   letter-spacing: -0.3px;
@@ -24,9 +23,8 @@
 }
 
 .frame .div {
-  color: $color-system-gray-gray80;
+  color: var(--color-system-gray-gray80);
   flex: 1;
-  font-family: 'Apple SD Gothic Neo-Regular', Helvetica;
   font-size: 13px;
   font-weight: 400;
   letter-spacing: -0.3px;
